boutons

+++ db/migrate/20240220170043_create_cards.rb
@@ -3,9 +3,9 @@ class CreateCards < ActiveRecord::Migration[7.1]
     create_table :cards do |t|
       t.string :name
       t.text :description
-      t.float :damages
+      t.integer :damages
       t.text :image_card
-      t.float :pv
+      t.integer :pv
       t.text :action
       t.string :attack_name
